Updated: Command security and easy-of-use
This changes the way that checks work to tell if a user can use a command. It also checks to see if a message is a full fantasy command, rather than just if it starts with the character (this will send invalid fantasy commands to Minecraft). This is all untested.