Update patch set 1
Patch Set 1:

(1 comment)

I couldn't compile with the VanishNoPacket artifact still as a dependency, so I elected to drop VNP support because it seems that it's unmaintained; I can't find a single page related to it that's been touched in the last year. The code with which it was integrated is also something I expect to change quite a bit with refactoring, so it made sense to me to drop it; if we decide we need the integration, we can redo it later at the same effort (and probably better design).

Patch-set: 1
Reviewer: Gerrit User 5 <5@cf5d0eaf-4814-484b-b55f-41cd7f22928e>
Label: Verified=0