Request IRCv3 CAP chghost if available
This capability requests the IRCd send a CHGHOST message instead
of a faked QUIT/JOIN if a client changes their hostname.

Pircbotx currently doesn't support this message, but as DragonIRC
does precisely nothing with hostnames anyway it's fine to ignore it.
One side effect of enabling it is that "changing hostname" QUIT
messages aren't echoed to Minecraft at all.

I've tested this against InspIRCd2 which does *not* support CHGHOST,
and there is no behavioural differences. I've also tested this
against Solanum which *does* support CHGHOST, and the fake QUIT/JOIN
messages are hidden.

Change-Id: I10b690665838d21001ddfac85e9bfd62febe7f38